tests: consolidate FSM customization tests into single module
